+++ business/system/lib/system/repository/remote/sys_oper_log_api.dart
@@ -65,7 +65,7 @@ class SysOperLogRepository {
 
   ///删除操作日志
   static Future<IntensifyEntity<dynamic>> delete(CancelToken cancelToken,
-      {int? id, List<int>? ids}) {
+      {BigInt? id, List<BigInt>? ids}) {
     //参数校验
     assert(id != null && ids == null || id == null && ids != null);
     ids ??= [id!];
