+++ business/system/lib/system/ui/config/config_list_browser_page.dart
@@ -169,7 +169,7 @@ class _ConfigListBrowserVm extends AppBaseVm {
   }
 
   //删除事件
-  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<int>? idList}) {
+  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<BigInt>? idList}) {
     if (idList == null) {
       List<SysConfig> selectList = SelectUtils.getSelect(listVmSub.dataList) ?? [];
       if (selectList.isEmpty) {
@@ -177,7 +177,7 @@ class _ConfigListBrowserVm extends AppBaseVm {
         return;
       }
       List<String> nameList = selectList.map<String>((item) => item.configName!).toList();
-      List<int> idList = selectList.map<int>((item) => item.configId!).toList();
+      List<BigInt> idList = selectList.map<BigInt>((item) => item.configId!).toList();
       confirmHandler?.call(nameList).then((value) {
         if (value == true) {
           onDelete(idList: idList);
