+++ business/system/lib/system/ui/dict/data/dict_data_list_browser_page.dart
@@ -174,7 +174,7 @@ class _DictDataListBrowserVm extends AppBaseVm {
   }
 
   //删除事件
-  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<int>? idList}) {
+  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<BigInt>? idList}) {
     if (idList == null) {
       List<SysDictData> selectList = SelectUtils.getSelect(listVmSub.dataList) ?? [];
       if (selectList.isEmpty) {
@@ -182,7 +182,7 @@ class _DictDataListBrowserVm extends AppBaseVm {
         return;
       }
       List<String> nameList = selectList.map<String>((item) => item.dictLabel!).toList();
-      List<int> idList = selectList.map<int>((item) => item.dictCode!).toList();
+      List<BigInt> idList = selectList.map<BigInt>((item) => item.dictCode!).toList();
       confirmHandler?.call(nameList).then((value) {
         if (value == true) {
           onDelete(idList: idList);
