+++ business/system/lib/system/ui/menu/tree/menu_tree_page_vd.dart
@@ -67,7 +67,7 @@ class MenuTreePageWidget {
 class MenuTreeListDataBaseVmSub extends TreeFastBaseListDataVmSub<SysMenuTree> {
   final FastVm fastVm;
 
-  final List<int>? checkedIds;
+  final List<BigInt>? checkedIds;
 
   final SysMenu _currentMenuSearch = SysMenu();
 
@@ -200,7 +200,7 @@ class MenuTreeListDataBaseVmSub extends TreeFastBaseListDataVmSub<SysMenuTree> {
 
 ///角色菜单树数据VmSub
 class RoleMenuTreeListDataVmSub extends MenuTreeListDataBaseVmSub {
-  final int? roleId;
+  final BigInt? roleId;
 
   RoleMenuTreeListDataVmSub(super.fastVm, {this.roleId, super.checkedIds});
 
@@ -219,7 +219,7 @@ class RoleMenuTreeListDataVmSub extends MenuTreeListDataBaseVmSub {
 
 ///租户套餐菜单树数据VmSub
 class TenantPackageMenuTreeListDataVmSub extends MenuTreeListDataBaseVmSub {
-  final int? packageId;
+  final BigInt? packageId;
 
   TenantPackageMenuTreeListDataVmSub(super.fastVm, {this.packageId, super.checkedIds});
