+++ business/user/lib/user/ui/dept/dept_list_browser_page.dart
@@ -194,7 +194,7 @@ class _DeptListBrowserVm extends AppBaseVm {
   }
 
   //删除事件
-  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<int>? idList}) {
+  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<BigInt>? idList}) {
     if (idList == null) {
       List<DeptTree> selectList = SelectUtils.getSelect(listVmSub.dataList) ?? [];
       if (selectList.isEmpty) {
@@ -202,7 +202,7 @@ class _DeptListBrowserVm extends AppBaseVm {
         return;
       }
       List<String> nameList = selectList.map<String>((item) => item.label).toList();
-      List<int> idList = selectList.map<int>((item) => item.id!).toList();
+      List<BigInt> idList = selectList.map<BigInt>((item) => item.id!).toList();
       confirmHandler?.call(nameList).then((value) {
         if (value == true) {
           onDelete(idList: idList);
