+++ business/user/lib/user/ui/role/role_list_browser_page.dart
@@ -155,7 +155,7 @@ class _RoleListBrowserVm extends AppBaseVm {
   }
 
   //删除事件
-  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<int>? idList}) {
+  void onDelete({Future<bool?> Function(List<String>)? confirmHandler, List<BigInt>? idList}) {
     if (idList == null) {
       List<Role> selectList = SelectUtils.getSelect(listVmSub.dataList) ?? [];
       //移除超级管理员角色
@@ -167,7 +167,7 @@ class _RoleListBrowserVm extends AppBaseVm {
         return;
       }
       List<String> nameList = selectList.map<String>((item) => item.roleName!).toList();
-      List<int> idList = selectList.map<int>((item) => item.roleId!).toList();
+      List<BigInt> idList = selectList.map<BigInt>((item) => item.roleId!).toList();
       confirmHandler?.call(nameList).then((value) {
         if (value == true) {
           onDelete(idList: idList);
